Tests/LibWeb/Text/expected/hit_testing/absolute-positioned-child-of-wrapped-inline.txt

Overlay starts near button: true
Overlay height does not span previous lines: true
Point above button does not hit overlay: true
